refactor: ディレクトリ構成の変更
